refactor extract ActiveModsList into own component
